feat: display fetched apps and their positions
